+++ crates/stackable-operator/src/git_sync/mod.rs
@@ -1,0 +1,4 @@
+//! Modules for git-sync
+
+pub mod framework;
+pub mod spec;