+++ crates/stackable-operator/src/crd/listener/class/mod.rs
@@ -31,6 +31,21 @@ pub mod versioned {
     pub struct ListenerClassSpec {
         pub service_type: core_v1alpha1::ServiceType,
 
+        /// Configures whether a LoadBalancer service should also allocate node ports (like NodePort).
+        ///
+        /// Ignored unless serviceType is LoadBalancer.
+        // TODO: v1alpha2: Move into ServiceType::LoadBalancer
+        #[serde(default = "ListenerClassSpec::default_load_balancer_allocate_node_ports")]
+        pub load_balancer_allocate_node_ports: bool,
+
+        /// Configures a custom Service loadBalancerClass, which can be used to access secondary
+        /// load balancer controllers that are installed in the cluster, or to provision
+        /// custom addresses manually.
+        ///
+        /// Ignored unless serviceType is LoadBalancer.
+        // TODO: v1alpha2: Move into ServiceType::LoadBalancer
+        pub load_balancer_class: Option<String>,
+
         /// Annotations that should be added to the Service object.
         #[serde(default)]
         pub service_annotations: BTreeMap<String, String>,