+++ crates/stackable-operator/CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to this project will be documented in this file.
   - Each CRD-related struct and enum has been versioned. The initial version is `v1alpha1`.
   - The `static` authentication provider must now be imported using `r#static`.
   - Import are now more granular in general.
+- BREAKING: Update to `kube` to `1.0.0` and `k8s-openapi` to `0.25.0`.
+  Use k8s `1.33` for compilation ([#1037]).
 
 ### Fixed
 
@@ -20,6 +22,7 @@ All notable changes to this project will be documented in this file.
 [#968]: https://github.com/stackabletech/operator-rs/pull/968
 [#1025]: https://github.com/stackabletech/operator-rs/pull/1025
 [#1029]: https://github.com/stackabletech/operator-rs/pull/1029
+[#1037]: https://github.com/stackabletech/operator-rs/pull/1037
 
 ## [0.92.0] - 2025-04-14