+++ crates/stackable-operator/CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.93.0] - 2025-05-19
+
 ### Changed
 
 - BREAKING: Version common CRD structs and enums ([#968]).